fix(gesture): 平坦化遷移の送信条件 (26.909.29)

+++ src/test/kotlin/com/awabi2048/ccsystem/core/gesturegui/GestureGuiVirtualOrientationTest.kt
@@ -36,7 +36,7 @@ class GestureGuiVirtualOrientationTest {
     }
 
     private fun Vector3f.dotOf(x: Double, y: Double, z: Double): Double =
-        (this.x * x + this.y * y + this.z * z).toDouble()
+        this.x * x + this.y * y + this.z * z
 
     @Test
     fun `block quatは旧ブロック向きに一致する`() {
